GH-3: Fix race condition in the partitions test
Fixes GH-3 (https://github.com/spring-projects/spring-kafka/issues/3)

The default `OffsetResetStrategy` is `LATEST` for the `KafkaConsumer` independently of the `KafkaMessageListenerContainer` option.
Therefore the `Thread.sleep(1000)` might not be enough in the test to wait before publishing messages to the topic.
With the race condition the listener may be initialized in the latest offset bypassing the required messages to poll.

* Add `initialConsumersLatch` to wait until the first `Consumer.poll()` will finish, assuming the offset position has been updated already.
* Remove dangerous `Thread.sleep(1000)`
* Add `assert` for the `latch2` from the second listener to be sure that everything works as expected.