GH-84: Pause/Resume
Resolves #84

Slow consumers can cause a rebalance when not required.

Solution: 2 threads per consumer
 - beanName-kafka-consumer-n
 - beanName-kafka-listener-n

The first constantly polls Kafka; for each retrieved message then hands off
the message to the second thread (via a `ListenableFuture`).
If the future completes ok (`get()`, all is well and we move on to the next
message, if not, the consumer is paused and we continue to poll (heartbeat).

When the callback completes successfully, the consumer is awoken from the poll
and the next previously retrieved message is processed. When all such messages
are processed, the consumer is resumed.

If manual acks are in play, the ack is passed from the listener thread back to
the consumer thread via a `BlockingQueue` so the actual commit is done on the
right thread.

If the delayed execution fails, normally, the error is reported to the error
handler, the message is ack'd and we move on to the next as with a successful
completion.

However, to satisfy the use case reported by issue 84, if the exception is
assignable to the configured `pausedException` class, the consumer is
paused as with a timeout. In this case, after the next poll, the same
record is attempted to be delivered. Deliveries continue until successful
or some other exception is thrown.

Properties:

- `pauseWhenSlow` - enable pause/resume (default `true`)
- `pauseAfter` - how long to wait for the listener thread to return
- `pauseException` - an exception that will cause a pause

Polishing

Polishing - Use a BlockingQueue

In order to maintain a pipeline of work for the listener invoker,
hand ConsumerRecords from each poll to the listener via a blocking
queue.

Now, pause the consumer only if the blocking queue becomes full.

The queue size is configurable (default 1) so we pause when 2
ConsumerRecords have been retrieved.

Remove the `pauseForException` code - instead use a `RetryTemplate` in
the listener invoker configured to retry for an exception of choice.

Since retry generally uses a backoff - this will put backpressure on the
queue and the consumer will be paused.

There's a slight semantic change to `AckMode.BATCH` in that it now means
acks for currently processed records will be committed after each poll;
it may not be the complete batch.

Suspend if not pauseEnabled

Enhacements to GH-84

- guarantee that the container is effectively stopped when stop() returns;
- On topic and topic pattern subscriptions, clear pending messages and
flush acknowledgments;
- On topic and topic pattern subscriptions, use `partitionsAssigned` as trigger
for starting the invoker

Additional enhancements to GH-84

- Run autocommit containers on a single thread

Additional enhancements to GH-84

- only start the invoker if the container is running and we have partitions to listen

Additional changes to GH-84

Use `cancel` to stop the listener and interrupt the poll
Allow for asynchronous stop in `stop(Runnable)`
Simplify classes used by the tests

Reinstate isPauseEnabled Use

Interrupt thread explicitly and wait for a clean stop on the invoker

Create ContainerProperties

Move ContainerProperties to Ctor Argument

Add topic config to the properties.

Since the annotated listener is created by a factory, we have
to "clone" the properties.

Doc Polishing for ContainerProperties

Move ContainerProperties to a Top Level Class

Simple polishing