Fix race condition in test on offset commits
The `ConcurrentMessageListenerContainerTests#testDefinedPartitions()` relies on the fact that previous batches have been committed to Broker
and starts a new test-case for defined offset.
Since commits are happened on the separate Thread from the `stop()` and more over the `consumer.wakeup()` may discard the currently queued commit request,
to properly perform assertions we have to wait until commits are finished and only after that stop the container.

* Since `testDefinedPartitions()` doesn't use concurrency, move the test to the `KafkaMessageListenerContainerTests`
* Add `stopLatch` barriers before each `stop()` and pass barriers on their `commitSync()`