Add Spring Data projection support
This commit introduces a `ProjectingMessageConverter` that supports binding
`String` and `byte[]`-backed JSON strings to Spring Data Projection interfaces.
This allows very selective, and low-coupled bindings to data including
the lookup of values from multiple places inside the JSON document.
For example the following interface can be defined as message payload type:

```
interface SomeSample {

  @JsonPath({ "$.username", "$.user.name" })
  String getUsername();
}
```

Accessor methods will be used to lookup the property name as field in the
received JSON document by default.
The @JsonPath expression allows to customize the value lookup and even
to define multiple JSONPath expression to lookup values from multiple
places until an expression returns an actual value.

If the type we're supposed to unmarshal into is a plain class the
`ProjectingMessageConverter` delegates to the default `StringJsonMessageConverter`.

Added Spring Data and the transitively required Jayway JSONPath
library as optional build dependencies.

The `ProjectingMessageConverter` now delegates writes to
a `StringJsonMessageConverter`.

* Some code style polishing